Split CLI tests.
