+++ tests/test_rio_edit_info.py
@@ -458,3 +458,67 @@ def test_colorinterp_like_all(
             ColorInterp.green,
             ColorInterp.blue,
             ColorInterp.alpha]
+
+
+
+def test_all_callback_pass(data):
+    ctx = MockContext()
+    ctx.obj['like'] = {'transform': 'foo'}
+    assert all_handler(ctx, None, None) is None
+
+
+def test_all_callback(data):
+    ctx = MockContext()
+    ctx.obj['like'] = {'transform': 'foo'}
+    assert all_handler(ctx, None, True) == {'transform': 'foo'}
+
+
+def test_all_callback_None(data):
+    ctx = MockContext()
+    assert all_handler(ctx, None, None) is None
+
+
+def test_transform_callback_pass(data):
+    """Always return None if the value is None"""
+    ctx = MockContext()
+    ctx.obj['like'] = {'transform': 'foo'}
+    assert transform_handler(ctx, MockOption('transform'), None) is None
+
+
+def test_transform_callback_err(data):
+    ctx = MockContext()
+    ctx.obj['like'] = {'transform': 'foo'}
+    with pytest.raises(click.BadParameter):
+        transform_handler(ctx, MockOption('transform'), '?')
+
+
+def test_transform_callback(data):
+    ctx = MockContext()
+    ctx.obj['like'] = {'transform': 'foo'}
+    assert transform_handler(ctx, MockOption('transform'), 'like') == 'foo'
+
+
+def test_crs_callback_pass(data):
+    """Always return None if the value is None."""
+    ctx = MockContext()
+    ctx.obj['like'] = {'crs': 'foo'}
+    assert crs_handler(ctx, MockOption('crs'), None) is None
+
+
+def test_crs_callback(data):
+    ctx = MockContext()
+    ctx.obj['like'] = {'crs': 'foo'}
+    assert crs_handler(ctx, MockOption('crs'), 'like') == 'foo'
+
+
+def test_tags_callback_err(data):
+    ctx = MockContext()
+    ctx.obj['like'] = {'tags': {'foo': 'bar'}}
+    with pytest.raises(click.BadParameter):
+        tags_handler(ctx, MockOption('tags'), '?') == {'foo': 'bar'}
+
+
+def test_tags_callback(data):
+    ctx = MockContext()
+    ctx.obj['like'] = {'tags': {'foo': 'bar'}}
+    assert tags_handler(ctx, MockOption('tags'), 'like') == {'foo': 'bar'}
